django_enhanced/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_full_version <= '3.11.2'
autopep8==2.0.2 ; python_version >= '3.6'
backports.zoneinfo==0.2.1 ; python_version < '3.9'
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.2.2
django-cprofile-middleware==1.0.5
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-filter==23.2
django-redis==5.3.0
django-silk==5.0.3
djangorestframework==3.14.0
gprof2dot==2022.7.29 ; python_version >= '2.7'
httpie==3.2.2
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.7.0 ; python_version < '3.10'
markdown==3.4.3
markdown-it-py==3.0.0 ; python_version >= '3.8'
mdurl==0.1.2 ; python_version >= '3.7'
multidict==6.0.4 ; python_version >= '3.7'
pip==23.1.2 ; python_version >= '3.7'
psycopg2-binary==2.9.6
pycodestyle==2.10.0 ; python_version >= '3.6'
pygments==2.15.1 ; python_version >= '3.7'
pysocks==1.7.1
python-decouple==3.8
pytz==2023.3
redis==4.6.0 ; python_version >= '3.7'
requests[socks]==2.31.0 ; python_version >= '3.7'
requests-toolbelt==1.0.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rich==13.4.2 ; python_full_version >= '3.7.0'
setuptools==68.0.0 ; python_version >= '3.7'
snakeviz==2.2.0
sqlparse==0.4.4 ; python_version >= '3.5'
tomli==2.0.1 ; python_version < '3.11'
tornado==6.3.2 ; python_version >= '3.8'
typing-extensions==4.7.0 ; python_version < '3.11'
urllib3==2.0.3 ; python_version >= '3.7'
zipp==3.15.0 ; python_version >= '3.7'
